Migrating to webpack bundled code
